Mention that relative paths rely on cwd
It was not clear that relative paths rely on the cwd. For example, a valid alternative interpretation could be that relative paths rely on the location of the eslintconfig file.

This change clarifies how the `parserOptions.project` relative paths are interpreted, and ties it in nicely with the existing recommendation to look at the `tsconfigRootDir` setting.